services/analytics-python/requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
httpx==0.28.1
pydantic==2.12.5
pytest==9.0.2
respx==0.22.0
python-dotenv==1.2.2
